Implement next and prev buttons on Step Detail Fragment
To accomplish that, I had to set the Step list on Ingredient and Step Fragment so it can be accessed on Step Detail Fragment.
Once it's set on ViewModel, you get through the next or the previous step by accessing the current step id and sending to the model. You can also access the stepList size so the id doesn't fall on out of bounds exception error.