Fix playWhenReady state value cannot be restored if the device is rotated (Nexus 5X, API 27)
When device was rotated when playing a video, it would show a blank screen. My reviewer suggested to retrieve
the values onPause but it didn't work as expected: when you first rotated, the player picked up where it left as expected
and as you rotated back it would get restored as in the initial state before it gets rotated. So, I move the restoring values
to onResume. It works on Nexus 5X, API 27 as expected.